@@ -10,12 +10,12 @@
         }
     ],
     "require": {
-        "php": "^7.1.3",
+        "php": "^7.2",
         "symfony/var-dumper": "^3.4 || ^4.0 || ^5.0"
     },
     "require-dev": {
         "mockery/mockery": "^1.0",
-        "phpunit/phpunit": "^7.0",
+        "phpunit/phpunit": "^8.3",
         "nesbot/carbon": "^2.23.0"
     },
     "autoload": {
@@ -31,6 +31,7 @@
         "files": [
             "tests/files/Support/Carbon.php",
             "tests/files/Support/HtmlString.php",
+            "tests/files/Support/HigherOrderTapProxy.php",
             "tests/files/Support/Str.php",
             "tests/files/Support/Stringable.php"
         ]
