@@ -195,15 +195,19 @@ class Stringable
     }
 
     /**
-     * Split a string using a regular expression.
+     * Split a string using a regular expression or by length.
      *
-     * @param  string  $pattern
+     * @param  string|int  $pattern
      * @param  int  $limit
      * @param  int  $flags
      * @return \Illuminate\Support\Collection
      */
     public function split($pattern, $limit = -1, $flags = 0)
     {
+        if (filter_var($pattern, FILTER_VALIDATE_INT) !== false) {
+            return collect(mb_str_split($this->value, $pattern));
+        }
+
         $segments = preg_split($pattern, $this->value, $limit, $flags);
 
         return ! empty($segments) ? collect($segments) : collect();
